Auth Service: MVC refactor
