+++ .gitignore
@@ -1,3 +1,4 @@
 venv/
 __pycache__/
 .env
+__init__.py
