+++ src/auth_service/utils/logger.py
@@ -1,0 +1,7 @@
+import logging
+
+LOGGING = True
+logging.basicConfig(level=logging.INFO)
+logger = logging.getLogger(__name__)
+if LOGGING:
+    logger.warning("LOGGER ENABLED")
